pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "video-translator"
version = "0.1.0"
description = "Translate English videos to Spanish using ElevenLabs TTS"
requires-python = ">=3.11"
dependencies = [
    "elevenlabs==2.35.0",
    "openai==2.17.0",
    "click==8.2.1",
    "pydantic==2.12.5",
    "rich==14.3.2",
    "python-dotenv==1.1.0",
    "python-pptx>=1.0.0",
    "scikit-image>=0.22.0",
]

[project.scripts]
video-translate = "video_translator.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
